refactor(server): split skills-loader.js into 3 sibling modules (#3223)
The loader grew from 231 lines (pre-#3220) to ~1626 lines after the
trust pipeline (#3204), v2 frontmatter (#3197/3198/3199/3200), trust
hardening (#3215/3216/3218), per-tier budget (#3222), and several
related additions all landed. Single file mixed four distinct
concerns; targeted review at this size was getting hard.

Split per the issue's suggested topology:

- skills-frontmatter.js (371 lines)
  - parseFrontmatter (re-exported from skills-loader for back-compat)
  - YAML subset parser internals (_parseFrontmatterBody, _unquote,
    _stripUnquotedTrailingComment)
  - FRONTMATTER_KEYS, VALID_ACTIVATION_MODES, VALID_INJECTION_MODES
  - Provider gating (#3198): _skillMatchesProvider,
    _normalizeProviderName, _isClaudeFamilyProvider
  - Manual activation (#3199): _skillIsActive, _coerceManualSet
  - Injection mode (#3200): _resolveInjectionMode, _normalizeInjectionMode

- skills-content-validator.js (106 lines)
  - DEFAULT_ALLOWED_EXTENSIONS, SKIP_DIRECTORY_NAMES
  - _normalizeExtension (allowlist input coercion)
  - _bufferLooksLikeText (#3203 + #3216 full-content text sniff)
  - _pathLabel (#3215 sanitised log identifier)

- skills-budget.js (162 lines)
  - DEFAULT_MAX_SKILL_BYTES, DEFAULT_MAX_TOTAL_SKILL_BYTES,
    DEFAULT_SKILL_PRIORITY
  - _enforceTotalBudget + _priorityOf (#3202 priority-aware prune)
  - _filterByProviderAllowlist (#3207 operator-side scoping)

skills-loader.js (1087 lines, was 1626) keeps the orchestration:
loadActiveSkills, loadActiveSkillsLayered, findRepoSkillsDir,
formatSkillsForPrompt, groupSkillsByInjectionMode, findSkillForRetrust,
SKILLS_PROMPT_HEADER, DEFAULT_SKILLS_DIR.

Public API unchanged. Internal callers continue to import from
skills-loader.js — the helpers are re-exported where needed
(parseFrontmatter is the only external consumer; the rest are private).

Tests: 317 pass across skills-loader, settings-handlers, base-session,
skills-trust. Pure refactor — no logic changes.

Closes #3223
